requirements/install.txt
django
paver
django-extensions
django-click
